Main_version_1/requirements.txt
#Requirement for loading pdfs (install in root directory)
langchain 
openai 
python-dotenv 
langchain-community
pypdf
langchain_openai
sqlalchemy 
psycopg2-binary 
pydantic
